Qt UI overhaul, Windows import fix, and repo cleanup
UI (PyQt6 frontend):
- Processing page: remove preset card, default display levels to raw
  full-range (0-100%), collapsible left/right panels with persisted state
- B-Scan toolbar: zoom in/out, fit, 1:1 buttons with content-based width
- Screen adaptation: adaptive window size (min 92% of available geometry),
  HighDpi PassThrough rounding policy, compact tool rows so controls no
  longer clip at 125%/150% scaling or on small screens
- Hide trace_qc / equidistant_trace_resample from method browser
- Global font 10pt; log panel collapsed by default

Bug fixes:
- storage_primitives: use fsync_file() in atomic_output_path; fsync on a
  read-only handle fails on Windows (Errno 9) and broke every CSV import

Cleanup:
- Archive v0.8/v0.9-era release notes, implementation reports, audit docs
  and one-off scripts to archive/history_v0.9.x/
- Remove orphan modules (workflow_template_manager, export_performance),
  dead configs/ marker dir and broken run_quality_gate.bat
- Rewrite CLAUDE.md and AGENTS.md to current architecture

Verified: 719 passed / 1 pre-existing Windows failure; GUI smoke 8 shots OK

+++ PythonModule/sec_gain.py
@@ -27,7 +27,7 @@ def method_sec_gain(
     power = max(to_float(power, default=1.0), 1.0e-6)
 
     n_samples = int(arr.shape[0])
-    t = np.linspace(0.0, 1.0, n_samples, dtype=np.float32) ** np.float32(power)
+    t = np.linspace(0.0, 1.0, n_samples, dtype=np.float64) ** np.float32(power)
     gain_curve = (gain_min + (gain_max - gain_min) * t).astype(np.float32)
     result = arr * gain_curve[:, None]
